Added functionality to HealthService, updated HealthController to leverage HealthService

+++ HealthCheck/Models/Settings.cs
@@ -10,7 +10,7 @@ namespace HealthCheck.Models
         public string Url { get; set; }
         public string ValidationText { get; set; }
         public int Retries { get; set; }
-        public long Timeout { get; set; }
+        public int Timeout { get; set; }
         public bool ForceOnline { get; set; }
         public bool RequireValidSSL { get; set; }
     }
